+++ packages/api-client/src/event/index.ts
@@ -21,3 +21,4 @@ export * from './ConversationEvent';
 export * from './BackendEvent';
 export * from './TeamEvent';
 export * from './UserEvent';
+export * from './FederationEvent';