+++ packages/core/src/errors/FederatedBackendsError.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2019 Wire Swiss GmbH
+ * Copyright (C) 2023 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,11 +17,16 @@
  *
  */
 
-import {jsx as JSX} from '@emotion/react';
-import {create} from 'react-test-renderer';
+/**
+ * This error means we are trying to add users that are parts of 2 backends that are not federating with each other to a new conversation.
+ */
+export class NonFederatingBackendsError extends Error {
+  constructor(public readonly backends: string[]) {
+    super('2 backends are not connected');
+    this.name = 'NonFederatingBackendError';
+  }
+}
 
-import {StyledApp, THEME_ID} from '../Layout';
-
-/* eslint-disable no-unsanitized/method */
-export const matchComponent = (component: JSX.Element, themeId = THEME_ID.LIGHT) =>
-  (expect as any)(create(<StyledApp themeId={themeId}>{component}</StyledApp>).toJSON()).toMatchSnapshot();
+export function isNonFederatingBackendsError(error: unknown): error is NonFederatingBackendsError {
+  return !!error && typeof error === 'object' && 'name' in error && error.name === 'NonFederatingBackendError';
+}