+++ packages/react-ui-kit/src/Icon/CertificateExpired.tsx
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2019 Wire Swiss GmbH
+ * Copyright (C) 2023 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,11 +17,18 @@
  *
  */
 
-import {jsx as JSX} from '@emotion/react';
-import {create} from 'react-test-renderer';
+import {SVGIcon, SVGIconProps} from './SVGIcon';
 
-import {StyledApp, THEME_ID} from '../Layout';
-
-/* eslint-disable no-unsanitized/method */
-export const matchComponent = (component: JSX.Element, themeId = THEME_ID.LIGHT) =>
-  (expect as any)(create(<StyledApp themeId={themeId}>{component}</StyledApp>).toJSON()).toMatchSnapshot();
+export const CertificateExpiredIcon = (props: SVGIconProps) => (
+  <SVGIcon realWidth={16} realHeight={16} {...props}>
+    <path
+      d="M8.00661 0.778123L14.25 2.44775V8C14.25 11.5269 11.6778 14.3426 8.00089 15.2298C4.35861 14.3417 1.75 11.5234 1.75 8V2.56573L8.00661 0.778123Z"
+      stroke="#C20013"
+      strokeWidth="1.5"
+      strokeMiterlimit="16"
+      fill="transparent"
+    />
+    <rect x="7" y="3.5" width="2" height="5" fill="#C20013" />
+    <rect x="7" y="9.5" width="2" height="2" fill="#C20013" />
+  </SVGIcon>
+);