+++ packages/react-ui-kit/src/Icon/MLSVerified.tsx
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2019 Wire Swiss GmbH
+ * Copyright (C) 2023 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,11 +17,15 @@
  *
  */
 
-import {jsx as JSX} from '@emotion/react';
-import {create} from 'react-test-renderer';
+import {SVGIcon, SVGIconProps} from './SVGIcon';
 
-import {StyledApp, THEME_ID} from '../Layout';
-
-/* eslint-disable no-unsanitized/method */
-export const matchComponent = (component: JSX.Element, themeId = THEME_ID.LIGHT) =>
-  (expect as any)(create(<StyledApp themeId={themeId}>{component}</StyledApp>).toJSON()).toMatchSnapshot();
+export const MLSVerified = (props: SVGIconProps) => (
+  <SVGIcon realWidth={16} realHeight={16} {...props}>
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M15 8V1.87197L8 0L1 2V8C1 12 4.00718 15.0977 8 16C12.0344 15.0977 15 12 15 8ZM13 5.34572L6.85214 12L3 7.84605L4.23094 6.50033L6.85214 9.30856L11.7691 4L13 5.34572Z"
+      fill="#1D7833"
+    />
+  </SVGIcon>
+);