+++ packages/react-ui-kit/src/Icon/ProteusVerified.tsx
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2019 Wire Swiss GmbH
+ * Copyright (C) 2023 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,11 +17,21 @@
  *
  */
 
-import {jsx as JSX} from '@emotion/react';
-import {create} from 'react-test-renderer';
+import {SVGIcon, SVGIconProps} from './SVGIcon';
 
-import {StyledApp, THEME_ID} from '../Layout';
-
-/* eslint-disable no-unsanitized/method */
-export const matchComponent = (component: JSX.Element, themeId = THEME_ID.LIGHT) =>
-  (expect as any)(create(<StyledApp themeId={themeId}>{component}</StyledApp>).toJSON()).toMatchSnapshot();
+export const ProteusVerified = (props: SVGIconProps) => (
+  <SVGIcon realWidth={16} realHeight={16} {...props}>
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M15 1.87197V8C15 12 12.0344 15.0977 8 16C4.00718 15.0977 1 12 1 8V2L8 0L15 1.87197Z"
+      fill="#0552A0"
+    />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M8 16C4.00718 15.0977 1 12 1 8V2L8 0C8 0 8 13 8 16Z"
+      fill="#6AA4DE"
+    />
+  </SVGIcon>
+);